% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covsel.embed.R
\name{covsel.embed}
\alias{covsel.embed}
\title{covsel.embed}
\usage{
covsel.embed(
  covdata,
  pa,
  weights = NULL,
  force = NULL,
  algorithms = c("glm", "gam", "rf"),
  ncov = ceiling(log2(length(which(pa == 1)))),
  maxncov = 12,
  nthreads = detectCores()/2,
  seed = 12345
)
}
\arguments{
\item{covdata}{A data.frame containing continuous covariate values extracted at presence–absence ('pa') locations.}

\item{pa}{A numeric vector indicating species presences (1) and absences (0).}

\item{weights}{A numeric vector of weights corresponding to each value in 'pa' (same length as 'pa').}

\item{force}{An optional character vector specifying the name(s) of covariate(s) to be forced into the final set.}

\item{algorithms}{A character vector specifying the algorithm(s) to be used for the embedding procedure (options: "glm", "gam", "rf").}

\item{ncov}{An integer specifying the target number of covariates to include in the final set.}

\item{maxncov}{An integer specifying the maximum number of covariates allowed in the final set.}

\item{nthreads}{An integer specifying the number of cores to be used during parallel operations.}

\item{seed}{An integer specifying the random seed for reproducibility.}
}
\value{
A list with three components:
\enumerate{
  \item \code{covdata}: A data.frame containing the covariates selected after the regularization, penalization, and ranking procedures.
  \item \code{ranks_1}: A data.frame containing the individual ranks of all covariates for each target algorithm.
  \item \code{ranks_2}: A data.frame containing the final average ranks of the selected covariates.
}
}
\description{
Covariate selection with model-specific embedding (Step-2)
}
\examples{
library(covsel)
covdata<-data_covfilter
dim(covdata)
covdata_embed<-covsel.embed(covdata, pa=data_covsel$pa, algorithms=c('glm','gam','rf'), seed=12345)
dim(covdata_embed$covdata)
}
\author{
Antoine Adde (antoine.adde@eawag.ch)
}
