% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covsel.filter.R
\name{covsel.filter}
\alias{covsel.filter}
\title{covsel.filter}
\usage{
covsel.filter(
  covdata,
  pa,
  weights = NULL,
  force = NULL,
  corcut = 0.7,
  variables = NULL,
  categories = NULL
)
}
\arguments{
\item{covdata}{A data.frame containing continuous covariate values extracted at presence–absence ('pa') locations.}

\item{pa}{A numeric vector indicating species presences (1) and absences (0).}

\item{weights}{A numeric vector of weights corresponding to each value in 'pa' (same length as 'pa').}

\item{force}{An optional character vector specifying the name(s) of covariate(s) to be forced into the final set.}

\item{corcut}{A numeric value specifying the correlation coefficient threshold used to identify collinearity.}

\item{variables}{A character vector of length equal to ncol(covdata) containing variable-level information.}

\item{categories}{A character vector of length equal to ncol(covdata) containing category-level information.}
}
\value{
A data.frame containing the set of non-collinear candidate covariates.
}
\description{
Apply the collinearity filtering algorithm at each target level (i=variable level; ii=category level; iii= all remainders)
}
\examples{
library(covsel)
covdata<-data_covsel$env_vars
dim(covdata)
covdata_filter<-covsel.filter(covdata,
                              pa=data_covsel$pa,
                              variables=data_covsel$catvar$variable,
                              categories=data_covsel$catvar$category)
dim(covdata_filter)
}
\author{
Antoine Adde (antoine.adde@eawag.ch)
}
