% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_covfilter-data.R
\docType{data}
\name{data_covfilter}
\alias{data_covfilter}
\title{Example covsel.filter output data}
\format{
A data frame containing 37 filtered covariates at 13609 locations
}
\usage{
data(data_covfilter)
}
\description{
A ready-to-use output from the covsel.filter function
}
\author{
Antoine Adde, 2022-10-16
}
\keyword{datasets}
