% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_covsel-data.R
\docType{data}
\name{data_covsel}
\alias{data_covsel}
\title{Generic data for running the main covsel example}
\format{
A list containing a "pa" vector with 3609 species presences (1) and 10k background absences (0); and data frame with 75 candidate covariates at the 13609 locations
}
\usage{
data(data_covsel)
}
\description{
Ready-to-use input data for running covsel demonstration
}
\author{
Antoine Adde, 2022-10-16
}
\keyword{datasets}
