% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covsel.filter.R
\name{covsel.filter}
\alias{covsel.filter}
\title{covsel.filter}
\usage{
covsel.filter(
  covdata,
  pa,
  variables = NULL,
  categories = NULL,
  weights = NULL,
  force = NULL,
  corcut = 0.7
)
}
\arguments{
\item{covdata}{data.frame containing covariate data}

\item{pa}{vector of species presences (1) and absences (0)}

\item{variables}{character vector of length ncol(covdata) containing variable-level names}

\item{categories}{character vector of length ncol(covdata) containing category-level names}

\item{weights}{vector containing the weights for each value in 'pa' (of length 'pa')}

\item{force}{optional character vector indicating the name(s) of the covariate(s) to be forced in the final set}

\item{corcut}{The value (numeric) of the correlation coefficient threshold for identifying colinearity}
}
\value{
A data.frame of "non-colinear" candidate covariates
}
\description{
Apply the colinearity filtering algorithm at each target level (i=variable level; ii=category level; iii= all remainders)
}
\examples{
covdata<-data_covsel$env_vars
dim(covdata)
covdata_filter<-covsel.filter(covdata,
                              pa=data_covsel$pa,
                              variables=data_covsel$catvar$variable,
                              categories=data_covsel$catvar$category)
dim(covdata_filter)
}
\author{
Antoine Adde (antoine.adde@unil.ch)
}
