% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covsel.filteralgo.R
\name{covsel.filteralgo}
\alias{covsel.filteralgo}
\title{covsel.filteralgo}
\usage{
covsel.filteralgo(covdata, pa, weights = NULL, force = NULL, corcut = 0.7)
}
\arguments{
\item{covdata}{A data.frame containing continuous covariate values extracted at presence–absence ('pa') locations.}

\item{pa}{A numeric vector indicating species presences (1) and absences (0).}

\item{weights}{A numeric vector of weights corresponding to each value in 'pa' (same length as 'pa').}

\item{force}{An optional character vector specifying the name(s) of covariate(s) to be forced into the final set.}

\item{corcut}{A numeric value specifying the correlation coefficient threshold used to identify collinearity.}
}
\value{
A data.frame containing the set of non-collinear candidate covariates.
}
\description{
Collinearity filtering algorithm
}
\examples{
library(covsel)
covdata<-data_covsel$env_vars
dim(covdata)
covdata_filter<-covsel.filteralgo(covdata, pa=data_covsel$pa)
dim(covdata_filter)
}
\author{
Antoine Adde (antoine.adde@eawag.ch)
}
